+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add functions
+    - `Cryptography.EncryptedEnvelope.encryptBatch`
+    - `Cryptography.EncryptedEnvelope.decryptBatch`
+- Add `Cryptography.Secret` module
 
 ## 2.2.0 - 2026-06-09
 - Add `Cryptography.Symmetric` module with `generateSecret`, `encrypt`, and `decrypt`